requirements.txt
# aiohttp==3.9.3
# aiosignal==1.3.1
# annotated-types==0.6.0
# anyio==3.7.1
# appnope==0.1.4
# argon2-cffi==23.1.0
# argon2-cffi-bindings==21.2.0
# arrow==1.3.0
# asttokens==2.4.1
# async-lru==2.0.4
# async-timeout==4.0.3
# attrs==23.2.0
# Babel==2.14.0
# beautifulsoup4==4.12.3
# bleach==6.1.0
# certifi==2024.2.2
# cffi==1.16.0
# charset-normalizer==3.3.2
# click==8.1.7
# comm==0.2.1
# dataclasses-json==0.6.4
# debugpy==1.8.1
# decorator==5.1.1
# defusedxml==0.7.1
# exceptiongroup==1.2.0
# executing==2.0.1
# fastapi==0.109.2
# fastjsonschema==2.19.1
# fqdn==1.5.1
# frozenlist==1.4.1
# h11==0.14.0
# httpcore==1.0.4
# httpx==0.27.0
# ibm-cloud-sdk-core==3.18.0
# ibm-cos-sdk==2.11.0
# ibm-cos-sdk-core==2.11.0
# ibm-cos-sdk-s3transfer==2.11.0
# ibm-watson-machine-learning==1.0.333
# ibm-watsonx-ai==0.0.5
# idna==3.6
# importlib-metadata==7.0.1
# ipykernel==6.29.2
# ipython==8.18.1
# ipywidgets==8.1.2
# isoduration==20.11.0
# jedi==0.19.1
# Jinja2==3.1.3
# jmespath==0.10.0
# json5==0.9.17
# jsonpatch==1.33
# jsonpointer==2.4
# jsonschema==4.21.1
# jsonschema-specifications==2023.12.1
# jupyter==1.0.0
# jupyter-console==6.6.3
# jupyter-events==0.9.0
# jupyter-lsp==2.2.2
# jupyter_client==8.6.0
# jupyter_core==5.7.1
# jupyter_server==2.12.5
# jupyter_server_terminals==0.5.2
# jupyterlab==4.1.2
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.25.3
# jupyterlab_widgets==3.0.10
# langchain==0.0.345
# langchain-core==0.0.13
# langsmith==0.0.92
# lomond==0.3.3
# MarkupSafe==2.1.5
# marshmallow==3.20.2
# matplotlib-inline==0.1.6
# mistune==3.0.2
# multidict==6.0.5
# mypy-extensions==1.0.0
# nbclient==0.9.0
# nbconvert==7.16.1
# nbformat==5.9.2
# nest-asyncio==1.6.0
# notebook==7.1.0
# notebook_shim==0.2.4
# numpy==1.26.4
# overrides==7.7.0
# packaging==23.2
# pandas==1.5.3
# pandocfilters==1.5.1
# parso==0.8.3
# pexpect==4.9.0
# platformdirs==4.2.0
# prometheus_client==0.20.0
# prompt-toolkit==3.0.43
# psutil==5.9.8
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pycparser==2.21
# pydantic==1.10.10
# pydantic_core==2.16.2
# Pygments==2.17.2
# PyJWT==2.8.0
# python-dateutil==2.8.2
# python-dotenv==1.0.1
# python-json-logger==2.0.7
# pytz==2024.1
# PyYAML==6.0.1
# pyzmq==25.1.2
# qtconsole==5.5.1
# QtPy==2.4.1
# referencing==0.33.0
# requests==2.31.0
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rpds-py==0.18.0
# Send2Trash==1.8.2
# six==1.16.0
# sniffio==1.3.0
# soupsieve==2.5
# SQLAlchemy==2.0.27
# stack-data==0.6.3
# starlette==0.36.3
# tabulate==0.9.0
# tenacity==8.2.3
# terminado==0.18.0
# tinycss2==1.2.1
# tomli==2.0.1
# tornado==6.4
# traitlets==5.14.1
# types-python-dateutil==2.8.19.20240106
# typing-inspect==0.9.0
# typing_extensions==4.9.0
# uri-template==1.3.0
# urllib3==1.26.18
# uvicorn==0.27.1
# wcwidth==0.2.13
# webcolors==1.13
# webencodings==0.5.1
# websocket-client==1.7.0
# widgetsnbextension==4.0.10
# wxai-langchain==0.0.4
# yarl==1.9.4
# zipp==3.17.0
langchain==0.0.345 
ibm_watsonx_ai
ibm_watson_machine_learning
python-dotenv 
fastapi
uvicorn
pydantic==1.10.10